Adding a function to bring historical data into the present. This also changes the downsampler setting.
To use this function add ts-X to the m= path, where X is the time to shift the data forward, like "1d"

For example: m=avg:ts-1d:rate:jvm.gc.CollectionTime{name=ParNew,host=szbigkraig} would plot the previous days garbage collection metrics. This is not very useful for GC information, but any metric with repeatable patterns can benefit from an overlay of the previous days or weeks data.

The downsampler setting has been adjusted to make it less ambiguous. Instead of a simple "5min-avg" to apply a 5 minute average "ds-5min-avg" is required.

This should also allow for more customized functions to be added to the m parameter, like custom titles.